Refactor DDE code to use LPCWSTR for string handling
Updated DDE-related structs and function signatures to use LPCWSTR instead of LPTSTR for improved type safety and Unicode compatibility. Cleaned up DDE code by removing redundant comments and standardizing #endif usage. Fixed minor bugs in DDE command implementations and ensured consistent string handling throughout. Enabled DDE support by uncommenting #define USING_DDE.

+++ Application/AeSys/DdeSys.cpp
@@ -70,7 +70,7 @@ HDDEDATA dde::SysReqFormats(UINT wFmt, HSZ, HSZ hszItem) {
 HDDEDATA dde::SysReqProtocols(UINT, HSZ, HSZ hszItem) {
   static wchar_t sz[] = L"Execute Control 1";
 
-  return DdeCreateDataHandle(ServerInfo.dwInstance, (LPBYTE)sz, wcslen(sz) + 1, 0, hszItem, CF_TEXT, 0);
+  return DdeCreateDataHandle(ServerInfo.dwInstance, (LPBYTE)sz, static_cast<DWORD>(wcslen(sz) + 1), 0, hszItem, CF_TEXT, 0);
   // Return a DDE data handle to the data object containing the return data.
 }
 // Process a request for the topics list.  Typically the data is returned in CF_TEXT format as a tab
@@ -96,4 +96,4 @@ HDDEDATA dde::SysReqTopics(UINT wFmt, HSZ, HSZ hszItem) {
   // Return a DDE data handle to the data object containing the return data.
   return hData;
 }
-#endif  // USING_DDE
+#endif
